Restricted Pages

+++ base/views.py
@@ -6,6 +6,7 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from .forms import RoomForm
 from django.contrib.auth.models import User
+from django.http import HttpResponse
 # Create your views here.
 
 rooms = [
@@ -15,6 +16,10 @@ rooms = [
 ]
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -60,7 +65,7 @@ def room(request, pk):
 
     
 
-@login_required(login_url='/login')
+@login_required(Login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -72,10 +77,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(Login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -86,8 +95,11 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(Login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
